% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestridge-package.R
\name{gravier}
\alias{gravier}
\title{breast cancer data set}
\format{
A list containing the design matrix X and response matrix y
}
\source{
\url{https://github.com/ramhiser}
}
\value{
No return value
}
\description{
Gravier et al. (2010) have considered small, invasive ductal carcinomas
without axillary lymph node involvement (T1T2N0) to predict metastasis of
small node-negative breast carcinoma. Using comparative genomic
hybridization arrays, they examined 168 patients over a five-year period.
The 111 patients with no event after diagnosis were labelled good, and the
57 patients with early metastasis were labelled poor.
}
\references{
Eleonore Gravier., Gaelle Pierron., and Anne Vincent-Salomon.
(2010). A prognostic DNA signature for T1T2 node-negative breast cancer
patients.
}
\keyword{datasets}
