% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patient.data}
\alias{patient.data}
\title{Lymphoma patients data set}
\format{
patient.data A list with survival times, staus and covariates from patients.
}
\description{
Lymphoma patients data set
}
\details{
A subset of the data set of lymphoma patients used in the study of Alizadeh et al. (2000) and also Simon et al.
(2011).
}
\references{
Alizadeh, A. A., et al. (2000). Distinct types of diffuse large B-cell lymphoma identified by gene
expression profiling. Nature, 403(6769), p.503
Simon, N., Friedman, J., Hastie, T., & Tibshirani, R. (2011). Regularization paths for Cox's proportional hazards model via coordinate descent. Journal of statistical software, 39(5), 1.
}
