% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.R
\name{recover}
\alias{recover}
\title{recover the coefficients from screening}
\usage{
recover(object, sparse = TRUE)
}
\arguments{
\item{object}{A "\code{bess}" object.}

\item{sparse}{Logical or NULL, specifying whether the coefficients should be
presented as sparse matrix or not.}
}
\description{
This function recovers the original coefficients in
the \code{beta.all} of the \code{bess} object when screening is utilized in \code{bess} function.
}
\details{
When screening is used, the active set updates are restricted on a chosen subset and the \code{beta.all}
 component of \code{bess} object contains only coefficients on this subset. This function helps recover
 the whole coefficients.
}
\examples{

# Generate simulated data
n <- 200
p <- 20
k <- 5
rho <- 0.4
SNR <- 10
cortype <- 1
seed <- 10
Data <- gen.data(n, p, k, rho, family = "gaussian", cortype = cortype, snr = SNR, seed = seed)
lm.bss <- bess(Data$x, Data$y, method = "sequential")

recover(lm.bss)
recover(lm.bss, best.model = FALSE)

}
\references{
Wen, C., Zhang, A., Quan, S. and Wang, X. (2020). BeSS: An R
Package for Best Subset Selection in Linear, Logistic and Cox Proportional
Hazards Models, \emph{Journal of Statistical Software}, Vol. 94(4).
doi:10.18637/jss.v094.i04.
}
\seealso{
\code{\link{bess}}.
}
\author{
Canhong Wen, Aijun Zhang, Shijie Quan, Liyuan Hu, Kangkang Jiang, Yanhang Zhang, Jin Zhu and Xueqin Wang.
}
