% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestridge-package.R
\name{SAheart}
\alias{SAheart}
\title{Risk factors associated with heart disease}
\format{
The variables are as follows:
\itemize{
\item sbp: Systolic blood pressure \item tobacco: Cumulative tobacco
consumption, in kg \item ldl: Low-density lipoprotein cholesterol \item
adiposity: Adipose tissue concentration \item famhist: Family history of
heart disease (1=Present, 0=Absent) \item typea: Score on test designed to
measure type-A behavior \item obesity: Obesity \item alcohol: Current
consumption of alcohol \item age: Age of subject \item chd: Coronary heart
disease at baseline; 1=Yes 0=No}

A data frame with 462 observations on 10 variables
}
\value{
No return value
}
\description{
Data from a subset of the Coronary Risk-Factor Study baseline survey,
carried out in rural South Africa.
}
\references{
Rousseauw, J., du Plessis, J., Benade, A., Jordaan, P., Kotze,
J. and Ferreira, J. (1983).  Coronary risk factor screening in three rural
communities. South African Medical Journal 64: 430-436.
}
\keyword{datasets}
